+++ test/root-descriptions/RootDescription.mock.ts
@@ -55,7 +55,7 @@ export const ROOT_DESCRIPTION_REQUIRED_PROPERTIES =
     '    </device>' +
     '</root>';
 
-    export const ROOT_DESCRIPTION_DEFAULT_HTTPS_PORT =
+export const ROOT_DESCRIPTION_DEFAULT_HTTPS_PORT =
     '<?xml version="1.0"?>' +
     '<root xmlns="urn:schemas-upnp-org:device-1-0">' +
     '    <specVersion>' +
@@ -136,6 +136,37 @@ export const ROOT_DESCRIPTION_WITH_LOWERCASE_MACADDRESS =
     '        <modelNumber>M1014</modelNumber>' +
     '        <modelURL>http://www.axis.com/</modelURL>' +
     '        <serialNumber>accc8e270ad8</serialNumber>' +
+    '        <UDN>uuid:Upnp-BasicDevice-1_0-accc8e270ad8</UDN>' +
+    '        <serviceList>' +
+    '            <service>' +
+    '                <serviceType>urn:axis-com:service:BasicService:1</serviceType>' +
+    '                <serviceId>urn:axis-com:serviceId:BasicServiceId</serviceId>' +
+    '                <controlURL>/upnp/control/BasicServiceId</controlURL>' +
+    '                <eventSubURL>/upnp/event/BasicServiceId</eventSubURL>' +
+    '                <SCPDURL>/scpd_basic.xml</SCPDURL>' +
+    '            </service>' +
+    '        </serviceList>' +
+    '        <presentationURL>http://192.168.1.102:80/</presentationURL>' +
+    '    </device>' +
+    '    <URLBase>http://192.168.1.102:51578/</URLBase>' +
+    '</root>';
+
+export const ROOT_DESCRIPTION_WITHOUT_MACADDRESS =
+    '<?xml version="1.0"?>' +
+    '<root xmlns="urn:schemas-upnp-org:device-1-0">' +
+    '    <specVersion>' +
+    '        <major>1</major>' +
+    '        <minor>0</minor>' +
+    '    </specVersion>' +
+    '    <device>' +
+    '        <deviceType>urn:schemas-upnp-org:device:Basic:1</deviceType>' +
+    '        <friendlyName>AXIS M1014 - accc8e270ad8</friendlyName>' +
+    '        <manufacturer>AXIS</manufacturer>' +
+    '        <manufacturerURL>http://www.axis.com/</manufacturerURL>' +
+    '        <modelDescription>AXIS M1014 Fixed Network Camera</modelDescription>' +
+    '        <modelName>AXIS M1014</modelName>' +
+    '        <modelNumber>M1014</modelNumber>' +
+    '        <modelURL>http://www.axis.com/</modelURL>' +
     '        <UDN>uuid:Upnp-BasicDevice-1_0-accc8e270ad8</UDN>' +
     '        <serviceList>' +
     '            <service>' +